# backend/src/infrastructure/webserver/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OmniVerse Nexus API",
    "version": "1.0.0",
    "description": "API documentation for the OmniVerse Nexus Enterprise Digital Twin Civilization Simulation Platform"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "name": { "type": "string" },
                  "organizationName": { "type": "string" },
                  "roleName": { "type": "string" }
                },
                "required": ["email", "password", "name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Invalid payload or user exists" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate user and return JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/projects": {
      "get": {
        "summary": "List all projects / digital twins",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "summary": "Create a new project",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "summary": "Retrieve project details with telemetry",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/api/projects/{id}/telemetry": {
      "post": {
        "summary": "Append telemetry log to digital twin project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nodeName": { "type": "string" },
                  "status": { "type": "string" },
                  "health": { "type": "number" },
                  "latency": { "type": "integer" },
                  "traffic": { "type": "number" },
                  "aqi": { "type": "integer" },
                  "energy": { "type": "number" }
                },
                "required": ["nodeName"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Telemetry logged" }
        }
      }
    },
    "/api/scaling-events": {
      "get": {
        "summary": "Get Kubernetes scaling logs",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Success"}
        }
      },
      "post": {
        "summary": "Log a scaling event action",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "serviceName": {"type": "string"},
                  "actionType": {"type": "string"},
                  "count": {"type": "integer"}
                },
                "required": ["serviceName", "actionType", "count"]
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Logged successfully"}
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "List all users in organization",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Success"}
        }
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "Get current authenticated user profile",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Success"}
        }
      },
      "patch": {
        "summary": "Update current user name or email",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {"type": "string"},
                  "email": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Updated"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}